Fixed svg icons on windows
